fix(nc-gui): some cells self review changes
